api v2 | app start from sub paths | separate test server | promiced login
