+++ package.json
@@ -5,8 +5,10 @@
   "main": "index.js",
   "scripts": {
     "lint": "npx eslint .",
-    "build": "npx eslint . && node ./dev/compile_templates.js && node server/server.js",
-    "start": "node server/server.js"
+    "build": "npx eslint . && node ./dev/compile_templates.js && npm run start",
+    "start": "concurrently --kill-others \"npm run start-static\" \"npm run start-api\"",
+    "start-static": "node server/server.js",
+    "start-api": "node server/server-api.js"
   },
   "keywords": [],
   "author": "Denis, Svetlana, Yuriy, Alexey",
@@ -16,6 +18,7 @@
     "pug": "^3.0.2"
   },
   "devDependencies": {
+    "concurrently": "^6.3.0",
     "eslint": "^7.32.0",
     "eslint-config-google": "^0.14.0",
     "eslint-plugin-import": "^2.24.2",
