+++ public/components/signupForm.js
@@ -38,13 +38,29 @@ export default function signupForm(showRegister) {
   // поля формы
   let emailInput = null;
   let passwordRepeatInput = null;
-  const loginInput = createInput('login', 'Логин', 'login',
+  const loginInput = createInput('text', 'Логин', 'login',
       '4-20 символов, первый символ - буква', true);
   const passwordInput = createInput('password', 'Пароль', 'password',
-      'Больше 8 символов. '+
-      'Используйте строчные и заглавные буквы, цифры',
+      'Больше 8 любых непробельных символов.',
       true,
   );
+  const passwordInputWrapper = document.createElement('div');
+  passwordInputWrapper.className = 'password';
+  const passwordShowBtn = document.createElement('a');
+  passwordShowBtn.className = 'password-control-btn';
+  passwordShowBtn.addEventListener('click', ({target}) => {
+    const input = passwordInputWrapper.querySelector('#input-password');
+    if (input.getAttribute('type') === 'password') {
+      target.classList.add('password-show');
+      input.setAttribute('type', 'text');
+    } else {
+      target.classList.remove('password-show');
+      input.setAttribute('type', 'password');
+    }
+  });
+  passwordInputWrapper.appendChild(passwordInput);
+  passwordInputWrapper.appendChild(passwordShowBtn);
+
 
   if (showRegister) {
     emailInput = createInput('email', 'e-mail', 'email', null, true);
@@ -59,23 +75,25 @@ export default function signupForm(showRegister) {
 
   // html5-свойства полей
   // https://habr.com/ru/post/545150/  - про регулярки
-  loginInput.autofocus = true;
+  // loginInput.autofocus = true;
   loginInput.pattern = '^[a-zA-Z][a-zA-Z0-9_]{4,20}$';
 
   passwordInput.pattern =
-    '^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z])(?=(.*[a-zA-Z]){4}).{8,256}$';
+    '^\\S{8,40}$';
+  // '^[A-Za-z0-9.,\\/#!$%\\^&\\*;:{}=\\-_`~()]{8,40}$';
+  // '^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z])(?=(.*[a-zA-Z]){4}).{8,256}$';
 
   if (showRegister) {
     emailInput.maxLength = 100;
     passwordRepeatInput.maxLength = 8;
-    passwordRepeatInput.maxLength = 256;
+    passwordRepeatInput.maxLength = 40;
   }
 
   form.appendChild(createInputRow(loginInput));
   if (showRegister) {
     form.appendChild(createInputRow(emailInput));
   }
-  form.appendChild(createInputRow(passwordInput));
+  form.appendChild(createInputRow(passwordInputWrapper));
   if (showRegister) {
     form.appendChild(createInputRow(passwordRepeatInput));
   }
@@ -91,7 +109,7 @@ export default function signupForm(showRegister) {
       ),
   );
   submitBtn.classList.add('submit-btn-row');
-  submitBtn.querySelector('#submitBtn').className = 'form-submit-btn';
+  submitBtn.querySelector('#input-submitBtn').className = 'form-submit-btn';
   form.appendChild(submitBtn);
 
   // submit action
@@ -116,15 +134,12 @@ export default function signupForm(showRegister) {
     }
 
     Ajax.post({
-      url: showRegister ? '/signup' : '/login',
+      url: showRegister ? '/user/signup' : '/user/login',
       // TODO: encryption
       body: {
         login,
         email,
         password,
-        // TODO:
-        // showRegister ? name,
-        // showRegister ? surname,
       },
     })
         .then(
