+++ public/flux/reducers.js
@@ -5,27 +5,17 @@ import {authorizationTypes,
   modalTypes,
 } from './types.js';
 
-// TODO: куда-нибудь перенести
-const headerLinksOnLogin = [
-  {name: 'Профиль', section: 'profilePage', href: '/profile'},
-  {section: 'logout', name: 'Выход'},
-];
-const headerLinksOnLogout = [
-  {name: 'Зарегистрироваться', section: 'signupModal'},
-  {name: 'Войти', section: 'loginModal'},
-];
-const sideBarLinks = ['hello'];
+import {
+  InitialMainPageState,
+  headerLinksOnLogout,
+  headerLinksOnLogin,
+} from '../models/mainPage.js';
+
+import {InitialUserState} from '../models/authorization.js';
+import {InitialPageState} from '../models/page.js';
+
 
 // TODO: function themeReducer(state = initialThemeState)
-
-
-const InitialUserState = {
-  login: '',
-  name: 'UNAUTHORIZED',
-  surname: 'UNAUTHORIZED',
-  email: 'UNAUTHORIZED',
-  score: 0,
-};
 
 /**
  * @param {Object} state
@@ -41,12 +31,6 @@ export function authorizeReducer(state = InitialUserState, action) {
   }
   return state;
 }
-
-
-const InitialPageState = {
-  page: 'main',
-  docTitle: 'SaberProject',
-};
 
 /**
  * @param {Object} state
@@ -91,19 +75,6 @@ export function signupFormReducer(state = InitialSignupFormState, action) {
   return state;
 }
 
-
-const InitialMainPageState = {
-  isAuthenticated: false,
-  trackedCardId: 'loading-card', // отслеживаемая запись в ленте для подгрузки
-  isLoading: false,              // отправлен ли запрос на сервер
-  idLastLoaded: '',              // запоминаем последнюю загруженную запись
-  lastScrollPos: 0,              // скрол для возврата к той же записи
-  login: '',                     // для какого пользователя подборка
-  cards: [],                     // массив загруженных новостей
-  doNotUpload: false,
-  headerLinks: headerLinksOnLogout,
-  sideBarLinks,
-};
 
 /**
  * @param {Object} state
