+++ public/models/authorization.js
@@ -1,0 +1,13 @@
+// ///////////////////////////////// //
+//
+//         App Configuration
+//
+// ///////////////////////////////// //
+
+export const InitialUserState = {
+  login: '',
+  name: 'UNAUTHORIZED',
+  surname: 'UNAUTHORIZED',
+  email: 'UNAUTHORIZED',
+  score: 0,
+};
