+++ public/models/page.js
@@ -1,0 +1,10 @@
+// ///////////////////////////////// //
+//
+//         App Configuration
+//
+// ///////////////////////////////// //
+
+export const InitialPageState = {
+  page: 'main',
+  docTitle: 'SaberProject',
+};
