+++ public/modules/ajax.js
@@ -1,5 +1,5 @@
 // const APIurl = 'http://87.228.2.178:8081';
-const APIurl = 'http://localhost:8080';
+const APIurl = 'http://localhost:8081/api/v1';
 
 /**
  * Поддерживаемые методы: GET и POST
